Created a skeleton for HoughTransformer.
I have not yet decided if I want to implement ImageProcess.
I would need a few helper methods to extract where the circle
is; which makes it atypical compared to the other ImageProcesses.
However, the advantage is that it can just be added to the ProcessList
and run through in a pipeline like the rest.